+++ CHANGELOG.md
@@ -13,6 +13,7 @@ and this project adheres to
 - ✨(backend) allow masking documents from the list view #1171
 - ✨(frontend) subdocs can manage link reach #1190
 - ✨(frontend) add duplicate action to doc tree #1175
+- ✨(frontend) add multi columns support for editor #1219
 
 ### Changed